Zerodha.Masterlist.Importer/requirements.txt
certifi==2022.6.15
charset-normalizer==2.1.0
greenlet==1.1.2
idna==3.3
PyMySQL==1.0.2
python-dotenv==0.20.0
requests==2.28.1
SQLAlchemy==1.4.39
urllib3==1.26.10